update principles controller

+++ app/controllers/principles_controller.rb
@@ -3,10 +3,15 @@
 class PrinciplesController < ApplicationController
   before_action :find_principle, only: %i[show destroy update]
   before_action :authorized, except: %i[login]
+  before_action :user_roles_list, except: %i[login]
 
   def index
-    @principles = Principle.all.limit(10).to_json({ include: { roles: { include: 'permissions' } } })
-    json_response(@principles, :created)
+    if current_user.class.name == 'Principle' || current_user.class.name == 'Super'
+      @principles = Principle.all.limit(10).to_json({ include: { roles: { include: 'permissions' } } })
+      json_response(@principles, :created)
+    else
+      json_response('Not authorized', :unauthorized)
+    end
   end
 
   def login
@@ -22,41 +27,61 @@ class PrinciplesController < ApplicationController
   end
 
   def create
-    @principle = Principle.create(principle_params)
-    @principle.password = random_password
-    if @principle.save
-      result = @principle.to_json({ include: { roles: { include: 'permissions' } } })
-      json_response(result, :created)
+    if current_user.class.name == 'Principle' || current_user.class.name == 'Super'
+      @principle = Principle.create(principle_params)
+      @principle.password = random_password
+      if @principle.save
+        result = @principle.to_json({ include: { roles: { include: 'permissions' } } })
+        json_response(result, :created)
+      else
+        json_response('Cannot create', :unprocessable_entity)
+      end
     else
-      json_response('Cannot create', :unprocessable_entity)
+      json_response('Invalid principle name or password', :unauthorized)
     end
   end
 
   def show
-    result = @principle.to_json({ include: { roles: { include: 'permissions' } } })
-    json_response(result, :created)
-  end
-
-  def update
-    if @principle.update(principle_params)
+    if current_user.class.name == 'Principle' || current_user.class.name == 'Super'
       result = @principle.to_json({ include: { roles: { include: 'permissions' } } })
       json_response(result, :created)
     else
-      json_response('Update errors', :unprocessable_entity)
+      json_response('Invalid principle name or password', :unauthorized)
+    end
+  end
+
+  def update
+    if current_user.class.name == 'Principle' || current_user.class.name == 'Super'
+      if @principle.update(principle_params)
+        result = @principle.to_json({ include: { roles: { include: 'permissions' } } })
+        json_response(result, :created)
+      else
+        json_response('Update errors', :unprocessable_entity)
+      end
+    else
+      json_response('Invalid principle name or password', :unauthorized)
     end
   end
 
   def destroy
-    if @principle.destroy
-      data = { 'message' => 'principle destroyed' }
-      json_response(data, :no_content)
+    if current_user.class.name == 'Principle' || current_user.class.name == 'Super'
+      if @principle.destroy
+        data = { 'message' => 'principle destroyed' }
+        json_response(data, :no_content)
+      else
+        data = { 'message' => 'Something went wrong' }
+        json_response(data, :no_content)
+      end
     else
-      data = { 'message' => 'Something went wrong' }
-      json_response(data, :no_content)
+      json_response('Invalid principle name or password', :unauthorized)
     end
   end
 
   private
+
+  def user_roles_list
+    determine_type(current_user)
+  end
 
   def find_principle
     @principle = Principle.find(params[:id])
@@ -68,3 +93,6 @@ class PrinciplesController < ApplicationController
     params.permit(:first_name, :last_name, :email, :password, role_ids: [])
   end
 end
+
+# "email" : "justine@gmail.com",
+# "password": "lNCvaUbm0A"
